Froze unselected tabs on ios
For primary tabs, ensured there's a default navigation controller because it isn't created until the children render. This doesn't happen straight away for the frozen tabs. Then updated the tabBarControllers view controllers when the navigation controller changes. Just done this in didSetProps for now - to test - but will probably signal from TabBarItem to TabBarView to update when child navigation controller comes in