Prevented frozen content from having display false
React sets frozen content to display false but this doesn't work for tab content. It changes the scroll position in the Twitter sample when switching between tabs. So created a native tab bar item container that can freeze and prevent display false - used same trick as already used on Scene