Updated view on create
First time in the content is null because it's frozen. So need to read the latest content when create is called again - this happens after switching to the tab. It's too late to fix it for first switch but it does work for subsequent switches - because it's now already updated from first switch