Showed content after first tab switch
For now, just to test, added the fragment async. This ensures the tab bar item child content has been added (from unfreeze). The async func has already run (when the current tab prop is set first) so need to refresh the fragment by removing and readding - this causes the create view of TabFragment to be called again!