Set current fragment sync instead of async
Thought the timing meant it had to be delayed so the child tab bar item was loaded before fragment. But it actually already was but just the setCurrentTab is called twice - the first time is too early. So just needed to make sure the fragment is reloaded the second time